+++ user_portal/__main__.py
@@ -5,7 +5,7 @@ from dotenv import load_dotenv
 
 load_dotenv()
 uvicorn.run(
-    "user_portal.app:create_app",
+    "server.entrypoints:users",
     factory=True,
     host=os.environ.get("USER_HOST", "127.0.0.1"),
     port=int(os.environ.get("USER_PORT", "3002")),